More robust argument checking, plus README update.
The `Validatable::validate` class method now not only guards against the
`validation_type` argument being an empty hash, but also makes sure that
`validation_type` is in fact a hash as well. In either event, an
exception is raised and a useful error message is displayed. Updated
relevant spec.

The README now describes validation syntax correctly and a little more
explicitly.

Closes #106.